Test: 기능 테스트, 예외 테스트 분리
